Fix / move tests for decorator & utils updates
