+++ neurodsp/tests/test_utils_decorators.py
@@ -1,6 +1,6 @@
 """Tests for sim decorator functions."""
 
-from neurodsp.sim.decorators import *
+from neurodsp.utils.decorators import *
 
 ###################################################################################################
 ###################################################################################################
